Modularize performance monitor helpers
