+++ src/core/performance/monitoring/performance_alerts.py
@@ -1,0 +1,43 @@
+#!/usr/bin/env python3
+"""Alerting utilities for performance monitoring."""
+
+from dataclasses import dataclass, field
+from enum import Enum
+from typing import Dict, Optional
+
+
+class AlertSeverity(Enum):
+    """Alert severity levels."""
+
+    INFO = "info"
+    WARNING = "warning"
+    CRITICAL = "critical"
+    EMERGENCY = "emergency"
+
+
+class AlertCondition(Enum):
+    """Conditions used for triggering alerts."""
+
+    GREATER_THAN = "greater_than"
+    LESS_THAN = "less_than"
+    EQUALS = "equals"
+    NOT_EQUALS = "not_equals"
+    GREATER_THAN_OR_EQUAL = "greater_than_or_equal"
+    LESS_THAN_OR_EQUAL = "less_than_or_equal"
+
+
+@dataclass
+class PerformanceAlert:
+    """Performance alert instance."""
+
+    alert_id: str
+    rule_name: str
+    metric_name: str
+    current_value: float
+    threshold: float
+    severity: AlertSeverity
+    message: str
+    timestamp: float
+    tags: Dict[str, str] = field(default_factory=dict)
+    resolved: bool = False
+    resolved_timestamp: Optional[float] = None
